+++ jumanji/environments/packing/flat_pack/generator.py
@@ -191,7 +191,7 @@ class RandomFlatPackGenerator(InstanceGenerator):
         arr_ = jnp.roll(arr_, -row_roll, axis=0)
         arr_ = jnp.roll(arr_, -col_roll, axis=1)
 
-        cropped_arr = jnp.zeros((3, 3), dtype=jnp.float32)
+        cropped_arr = jnp.zeros((3, 3), dtype=jnp.int32)
 
         cropped_arr = cropped_arr.at[:, :].set(arr_[:3, :3])
 
@@ -242,7 +242,7 @@ class RandomFlatPackGenerator(InstanceGenerator):
         col_interlock_idxs = get_significant_idxs(grid_col_dim)
 
         # Create an empty grid.
-        grid = jnp.ones((grid_row_dim, grid_col_dim))
+        grid = jnp.ones((grid_row_dim, grid_col_dim), dtype=jnp.int32)
 
         # Fill grid columns with block numbers
         (grid, _), _ = jax.lax.scan(
@@ -312,23 +312,23 @@ class ToyFlatPackGeneratorWithRotation(InstanceGenerator):
 
         solved_grid = jnp.array(
             [
-                [1.0, 1.0, 1.0, 2.0, 2.0],
-                [1.0, 1.0, 2.0, 2.0, 2.0],
-                [3.0, 1.0, 4.0, 4.0, 2.0],
-                [3.0, 3.0, 4.0, 4.0, 4.0],
-                [3.0, 3.0, 3.0, 4.0, 4.0],
+                [1, 1, 1, 2, 2],
+                [1, 1, 2, 2, 2],
+                [3, 1, 4, 4, 2],
+                [3, 3, 4, 4, 4],
+                [3, 3, 3, 4, 4],
             ],
-            dtype=jnp.float32,
+            dtype=jnp.int32,
         )
 
         blocks = jnp.array(
             [
-                [[0.0, 1.0, 0.0], [0.0, 1.0, 1.0], [1.0, 1.0, 1.0]],
-                [[2.0, 0.0, 0.0], [2.0, 2.0, 2.0], [2.0, 2.0, 0.0]],
-                [[0.0, 0.0, 3.0], [0.0, 3.0, 3.0], [3.0, 3.0, 3.0]],
-                [[4.0, 4.0, 0.0], [4.0, 4.0, 4.0], [0.0, 4.0, 4.0]],
+                [[0, 1, 0], [0, 1, 1], [1, 1, 1]],
+                [[2, 0, 0], [2, 2, 2], [2, 2, 0]],
+                [[0, 0, 3], [0, 3, 3], [3, 3, 3]],
+                [[4, 4, 0], [4, 4, 4], [0, 4, 4]],
             ],
-            dtype=jnp.float32,
+            dtype=jnp.int32,
         )
 
         return State(
